Refactor overlay warning screen and parental control dashboard
- Integrated PIN authentication in OverlayWarningScreen for adding more time.
- Improved code readability by consolidating widget properties and removing unnecessary line breaks.
- Updated the ParentDashboardScreen to include strict mode functionality and improved UI elements.
- Removed unused backspace handling in PinAuthScreen.
- Adjusted color scheme references in ScheduleScreen, AppCard, and UsageProgressBar for consistency.
- Changed DropdownButtonFormField to use initialValue instead of value in PrayerLockSettingsScreen.

+++ android/app/src/main/kotlin/com/example/mu_super_app/blocking/AppBlockingAccessibilityService.kt
@@ -74,7 +74,7 @@ class AppBlockingAccessibilityService : AccessibilityService() {
             val blockTimestamp = blockedApps[packageName] ?: return false
 
             val now = System.currentTimeMillis()
-            val blockDuration = prefs.getLong("${PFX}block_duration_minutes", 30) * 60 * 1000 // Convert to milliseconds
+            val blockDuration = prefs.getInt("${PFX}block_duration_minutes", 30).toLong() * 60 * 1000 // Convert to milliseconds
 
             // Check if block has expired
             if (now - blockTimestamp > blockDuration) {
@@ -144,7 +144,7 @@ class AppBlockingAccessibilityService : AccessibilityService() {
         val blockTimestamp = blockedApps[packageName] ?: return 0
 
         val now = System.currentTimeMillis()
-        val blockDuration = prefs.getLong("${PFX}block_duration_minutes", 30) * 60 * 1000
+        val blockDuration = prefs.getInt("${PFX}block_duration_minutes", 30).toLong() * 60 * 1000
         val elapsed = now - blockTimestamp
         val remaining = blockDuration - elapsed
